fix: change to use API from supabase

+++ share-hood/app/api/items/[id]/route.ts
@@ -1,18 +1,30 @@
 import {NextResponse} from "next/server";
 import {StatusCode} from "constants/statusCode";
-import {sql} from "@vercel/postgres";
+import sql from "@libs/db/db";
 import {ItemType} from "../../../../types/api/apiType";
+import {getLogger} from "@service/logger/logger";
 
 export async function GET(request: Request) {
   try {
     const {pathname} = new URL(request.url);
     const itemIDParam = pathname.split("/").pop();
+    if (!itemIDParam) {
+      return NextResponse.json(
+        {error: "Invalid item ID"},
+        {status: StatusCode.ERROR_BAD_REQUEST.code},
+      );
+    }
     const data = await sql<ItemType[]>`
-      SELECT * FROM "Item" WHERE "itemID" = ${itemIDParam};
+      SELECT * FROM "Item" WHERE "itemID" = ${itemIDParam}
     `;
-    return NextResponse.json(data.rows, {status: StatusCode.SUCCESS_OK.code});
+    getLogger("GET", request.url, "info", `data: ${JSON.stringify(data).replace(/"/g, " ")}`);
+
+    return NextResponse.json(data, {status: StatusCode.SUCCESS_OK.code});
   } catch (error: unknown) {
-    return NextResponse.json({error: error as string}, {status: StatusCode.ERROR_NOT_FOUND.code});
+    return NextResponse.json(
+      {error: error instanceof Error ? error.message : "An unknown error occurred"},
+      {status: StatusCode.ERROR_BAD_REQUEST.code},
+    );
   }
 }
 
@@ -34,26 +46,39 @@ export async function PUT(request: Request) {
       pickupLocation,
     }: ItemType = await request.json();
 
-    await sql`
-      UPDATE "Item"
-      SET
-        "userID" = ${userID},
-        "itemName" = ${itemName},
-        "itemDescription" = ${itemDescription},
-        "itemPrice" = ${itemPrice},
-        "itemQuantity" = ${itemQuantity},
-        "itemStatus" = ${itemStatus},
-        "category" = ${category},
-        "itemReturnDuration" = ${itemReturnDuration},
-        "dateAdded" = ${dateAdded},
-        "pickupLocation" = ${pickupLocation}
-      WHERE "itemID" = ${itemIDParam}
-    `;
+    if (!itemIDParam) {
+      return NextResponse.json(
+        {error: "Invalid item ID"},
+        {status: StatusCode.ERROR_BAD_REQUEST.code},
+      );
+    }
+
+    const data = await sql<ItemType[]>`
+        UPDATE "Item"
+        SET
+          "userID" = ${userID},
+          "itemName" = ${itemName},
+          "itemDescription" = ${itemDescription},
+          "itemPrice" = ${itemPrice},
+          "itemQuantity" = ${itemQuantity},
+          "itemStatus" = ${itemStatus},
+          "category" = ${category},
+          "itemReturnDuration" = ${itemReturnDuration},
+          "dateAdded" = ${dateAdded},
+          "pickupLocation" = ${pickupLocation}
+        WHERE "itemID" = ${itemIDParam}
+      `;
+
+    getLogger("PUT", request.url, "info", `data: ${JSON.stringify(data).replace(/"/g, " ")}`);
+
     return NextResponse.json(`Update Item: ${itemIDParam} Succeed`, {
       status: StatusCode.SUCCESS_OK.code,
     });
   } catch (error: unknown) {
-    return NextResponse.json({error: error as string}, {status: StatusCode.ERROR_BAD_REQUEST.code});
+    return NextResponse.json(
+      {error: error instanceof Error ? error.message : "An unknown error occurred"},
+      {status: StatusCode.ERROR_BAD_REQUEST.code},
+    );
   }
 }
 
@@ -62,14 +87,26 @@ export async function DELETE(request: Request) {
     const {pathname} = new URL(request.url);
     const itemIDParam = pathname.split("/").pop();
 
-    await sql`
-      DELETE FROM "Item"
-      WHERE "itemID" = ${itemIDParam}
-    `;
+    if (!itemIDParam) {
+      return NextResponse.json(
+        {error: "Invalid item ID"},
+        {status: StatusCode.ERROR_BAD_REQUEST.code},
+      );
+    }
+
+    const data = await sql<ItemType[]>`
+          DELETE FROM "Item"
+          WHERE "itemID" = ${itemIDParam}
+        `;
+
+    getLogger("DELETE", request.url, "info", `data: ${JSON.stringify(data).replace(/"/g, " ")}`);
     return NextResponse.json(`Delete Item: ${itemIDParam} Succeed`, {
-      status: StatusCode.SUCCESS_NO_CONTENT.code,
+      status: StatusCode.SUCCESS_CREATED.code,
     });
   } catch (error: unknown) {
-    return NextResponse.json({error: error as string}, {status: StatusCode.ERROR_BAD_REQUEST.code});
+    return NextResponse.json(
+      {error: error instanceof Error ? error.message : "An unknown error occurred"},
+      {status: StatusCode.ERROR_BAD_REQUEST.code},
+    );
   }
 }
